Generates schedules :)
program generates a list of modules which have an attached room/time/day
inside them.

All the database write  module should need to is read in an arraylist of
modules, take the id, room id, time and day (and convert this from in to
date) and place it inside a database.